requirements.txt
allure-pytest==2.14.0
allure-python-commons==2.14.0
attrs==25.3.0
certifi==2025.1.31
cffi==1.17.1
chromedriver-py==135.0.7049.84
colorama==0.4.6
faker==37.1.0
h11==0.14.0
idna==3.10
iniconfig==2.1.0
outcome==1.3.0.post0
packaging==24.2
pluggy==1.5.0
pycparser==2.22
PySocks==1.7.1
pytest==8.3.5
selenium==4.31.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.29.0
trio-websocket==0.12.2
typing-extensions==4.13.1
tzdata==2025.2
urllib3==2.3.0
websocket-client==1.8.0
wsproto==1.2.0
